fix: forgot append flag on one of the `tee` commands
man, what destructive mistake that was. with the `-a` flag, it just
overwrote the whole `/etc/fstab` config.
and on reboot it, the whole system would be read only and you wouldn't
be able to edit `fstab`
luckily you can use `sudo mount -o remount,rw /dev/root /` to make it.
writable again. using `cat /etc/mtab` on another similar system, i was
able to rebuild the `fstab` file 😞